+++ rust/src/apis.rs
@@ -1,12 +1,51 @@
+use crate::{error::Error, URL_PREFIX};
 use chrono::prelude::*;
+use reqwest::{header::HeaderMap, RequestBuilder, StatusCode};
+use serde::de::DeserializeOwned;
 
 pub mod get_business_comment_list;
 pub mod get_business_get;
 pub mod get_business_video_list;
 pub mod post_business_comment_reply_create;
 
-pub fn make_url(postfix: &str) -> String {
-    format!("{}{}", URL_PREFIX, postfix)
+const ENV_KEY: &str = "TICTOK_BUSINESS_PREFIX_API";
+
+pub fn clear_prefix_url() {
+    std::env::set_var(ENV_KEY, URL_PREFIX);
+}
+
+pub fn setup_prefix_url(url: &str) {
+    std::env::set_var(ENV_KEY, url);
+}
+
+pub(crate) fn make_url(postfix_url: &str, options: &Option<ApiOptions>) -> String {
+    make_url_with_prefix(
+        &std::env::var(ENV_KEY).unwrap_or(URL_PREFIX.to_owned()),
+        options,
+        postfix_url,
+    )
+}
+
+fn make_url_with_prefix(
+    default_perfix_url: &str,
+    options: &Option<ApiOptions>,
+    postfix_url: &str,
+) -> String {
+    let prefix_url = if let Some(options) = options {
+        if let Some(prefix_url) = options.prefix_url.as_ref() {
+            prefix_url
+        } else {
+            default_perfix_url
+        }
+    } else {
+        default_perfix_url
+    };
+    format!("{}{}", prefix_url, postfix_url)
+}
+
+#[derive(Debug, Clone, Default)]
+pub struct ApiOptions {
+    pub prefix_url: Option<String>,
 }
 
 #[derive(Debug)]
@@ -21,10 +60,6 @@ pub struct ResponseHeader {
     pub date: DateTime<Utc>,
     pub x_tt_log_id: String,
 }
-
-use crate::{error::Error, URL_PREFIX};
-use reqwest::{header::HeaderMap, RequestBuilder, StatusCode};
-use serde::de::DeserializeOwned;
 
 pub async fn execute_api<T>(builder: RequestBuilder) -> Result<ApiResponse<T>, Error>
 where